Simulation_Web/requirements.txt
Flask
pyserial
requests
cryptography